refactor: move campaigns routes to new folder structure
