+++ src/API/REST/V3/Routes/Campaigns/GetCampaignStatistics.php
@@ -1,15 +1,15 @@
 <?php
 
-namespace Give\Campaigns\Routes;
+namespace Give\API\REST\V3\Routes\Campaigns;
 
 use DateInterval;
 use DatePeriod;
 use DateTimeImmutable;
 use Exception;
+use Give\API\REST\V3\Routes\Campaigns\ValueObjects\CampaignRoute;
 use Give\API\RestRoute;
 use Give\Campaigns\CampaignDonationQuery;
 use Give\Campaigns\Models\Campaign;
-use Give\Campaigns\ValueObjects\CampaignRoute;
 use Give\Framework\Support\Facades\DateTime\Temporal;
 use WP_REST_Response;
 use WP_REST_Server;
